% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adherence_preprocess.R
\name{adherence_preprocess}
\alias{adherence_preprocess}
\title{Pre-process time data for adherence}
\usage{
adherence_preprocess(
  timestamps,
  regimen,
  patinfo = list(),
  nonmonit = data.frame()
)
}
\arguments{
\item{timestamps}{Input timestamps - vector of timestamps}

\item{regimen}{Regimen - regimen definition}

\item{patinfo}{Patient info - patient specific information}

\item{nonmonit}{Non-monitored date intervals}
}
\value{
A list of output variables
\itemize{
  \item \code{all_periods} - Processed timestamps into periods applying input parameters.
}
}
\description{
Pre-process time data for adherence
}
