% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncmap-package.R
\docType{package}
\name{oncmap-package}
\alias{oncmap}
\alias{oncmap-package}
\title{oncmap: Analyze Data from Electronic Adherence Monitoring Devices}
\description{
Medication adherence, defined as medication-taking behavior that aligns with the agreed-upon treatment protocol, is critical for realizing the benefits of prescription medications. Medication adherence can be assessed using electronic adherence monitoring devices (EAMDs), pill bottles or boxes that contain a computer chip that records the date and time of each opening (or “actuation”). Before researchers can use EAMD data, they must apply a series of decision rules to transform actuation data into adherence data. The purpose of this R package ('oncmap') is to transform EAMD actuations in the form of a raw .csv file, information about the patient, regimen, and non-monitored periods into two daily adherence values -- Dose Taken and Correct Dose Taken.
}
\author{
\strong{Maintainer}: Michal Kouril \email{Michal.Kouril@cchmc.org} (\href{https://orcid.org/0000-0002-4786-7934}{ORCID})

Authors:
\itemize{
  \item Meghan McGrady \email{Meghan.McGrady@cchmc.org} (\href{https://orcid.org/0000-0002-3150-3239}{ORCID})
  \item Mara Constance \email{Constance.Mara@cchmc.org} (\href{https://orcid.org/0000-0002-6776-8060}{ORCID})
  \item Kevin Hommel \email{Kevin.Hommel@cchmc.org} (\href{https://orcid.org/0000-0002-9913-509X}{ORCID})
}

}
\keyword{internal}
