% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_adherence.R
\name{report_adherence}
\alias{report_adherence}
\title{Report standarized output of the adherence processing}
\usage{
report_adherence(
  all_periods,
  timestamps,
  med,
  patinfo = list(),
  adhstart = NULL,
  adhend = NULL
)
}
\arguments{
\item{all_periods}{output of pre_adherence processing}

\item{timestamps}{timestamps dataframe from pre_adherence to calculate times and diffs in the report}

\item{med}{Medication name}

\item{patinfo}{Patient info - patient specific information}

\item{adhstart}{Report adherence start date}

\item{adhend}{Report adherence end date}
}
\description{
Report standarized output of the adherence processing
}
