% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input.R
\name{read_input}
\alias{read_input}
\title{Read input file}
\usage{
read_input(
  infile,
  include_formats = NULL,
  exclude_formats = NULL,
  formats_def = NULL,
  infile_data_output = FALSE
)
}
\arguments{
\item{infile}{Input CSV file name}

\item{include_formats}{Which formats to include in checking}

\item{exclude_formats}{Which formats to exclude from checking}

\item{formats_def}{New formats definition}

\item{infile_data_output}{Include infile data frame in the result}
}
\description{
Read input file
}
\examples{
# print(read_input("tests/testthat/ecap1.csv"))
# print(read_input("tests/testthat/ecap2.csv"))
# print(read_input("tests/testthat/simplemed.csv"))
# print(read_input("tests/testthat/mems.csv"))
# print(read_input("tests/testthat/adheretech.csv"))
# print(read_input("tests/testthat/ecap.csv"))
# print(read_input("tests/testthat/1077.csv"))
# print(read_input("tests/testthat/noformat.csv"))
# print(read_input("tests/testthat/nocsv.csv"))
}
